fix: :bug: atualização validações

+++ database/migrations/2022_12_29_010011_alter_google_color_add_field_created_by.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::table('google_colors', function (Blueprint $table) {
-            $table->integer('created_by')->unsigned()->nullable();
+            $table->bigInteger('created_by')->unsigned()->nullable();
             $table->foreign('created_by')->references('id')->on('users');
         });
     }
